Whoops, missed a file in last warning fix commit
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@9233 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/lib/codeclib.c
@@ -36,7 +36,7 @@ int codec_init(struct codec_api* rb)
 {
     local_rb = rb;
     mem_ptr = 0;
-    mallocbuf = (unsigned char *)rb->get_codec_memory((long *)&bufsize);
+    mallocbuf = (unsigned char *)rb->get_codec_memory((size_t *)&bufsize);
   
     return 0;
 }
